+++ src/main/zapHomeFiles/hud/display.js
@@ -99,6 +99,48 @@ Vue.component('dialog-modal', {
 	}
 });
 
+
+Vue.component('ajax-dialog-modal', {
+	template: '#ajax-dialog-modal-template',
+	props: ['show', 'title', 'text'],
+	methods: {
+		close: function() {
+			this.$emit('close');
+		},
+		buttonClick: function(id) {
+			this.port.postMessage({'action': 'dialogSelected', id: id, 'browserId': this.browser});
+			this.close();
+		}
+	},
+	data() {
+		return {
+			port: null,
+			browser: 'firefox-headless',
+			status: '',
+			buttons: [
+				{text: I18n.t("common_ok"), id:"okay"},
+				{text: I18n.t("common_cancel"), id:"cancel"}
+			]
+		}
+	},
+	created: function() {
+		let self = this;
+
+		eventBus.$on('showAjaxDialogModal', data => {
+			app.isAjaxDialogModalShown = true;
+			app.dialogModalTitle = data.title;
+			app.dialogModalText = data.text;
+
+			self.buttons = data.buttons;
+			self.port = data.port;
+			self.status = data.status;
+		})
+	},
+	beforeDestroy () {
+		eventBus.$off('showAjaxDialogModal')
+	}
+});
+
 Vue.component('select-tool-modal', {
 	template: '#select-tool-modal-template',
 	props:['show', 'title'],
@@ -294,6 +336,42 @@ Vue.component('simple-menu-modal', {
 	},
 	beforeDestroy () {
 		eventBus.$off('showSimpleMenuModal')
+	}
+})
+
+Vue.component('adv-menu-modal', {
+	template: '#adv-menu-modal-template',
+	props: ['show', 'title'],
+	methods: {
+		close: function() {
+			this.$emit('close');
+		},
+		itemSelect: function(itemId) {
+			this.port.postMessage({'action': 'itemSelected', 'id': itemId});
+			app.keepShowing = true;
+			app.isAdvMenuModalShown = false;
+			this.close();
+		}
+	},
+	data() {
+		return {
+			port: null,
+			items: {}
+		}
+	},
+	created() {
+		let self = this;
+
+		eventBus.$on('showAdvMenuModal', data => {
+			app.isAdvMenuModalShown = true;
+			app.iconMenuModalTitle = data.title;
+
+			self.items = data.items;
+			self.port = data.port;
+		})
+	},
+	beforeDestroy () {
+		eventBus.$off('showAdvMenuModal')
 	}
 })
 
@@ -786,6 +864,7 @@ document.addEventListener("DOMContentLoaded", () => {
 		el: '#app',
 		data: {
 			isDialogModalShown: false,
+			isAjaxDialogModalShown: false,
 			dialogModalTitle: "",
 			dialogModalText: "text",
 			isSelectToolModalShown: false,
@@ -797,6 +876,8 @@ document.addEventListener("DOMContentLoaded", () => {
 			alertDetailsModalTitle: I18n.t("alerts_details_title"),
 			isSimpleMenuModalShown: false,
 			simpleMenuModalTitle: I18n.t("common_menu_title"),
+			isAdvMenuModalShown: false,
+			advMenuModalTitle: I18n.t("common_menu_title"),
 			isBreakMessageModalShown: false,
 			breakMessageModalTitle: I18n.t("break_http_message_title"),
 			isHistoryMessageModalShown: false,
@@ -825,6 +906,21 @@ navigator.serviceWorker.addEventListener("message", event => {
 				title: config.title,
 				text: config.text,
 				buttons: config.buttons,
+				port: port
+			});
+
+			app.backStack.push(show)
+			show()
+
+			showDisplayFrame();
+			break;
+
+		case "showAjaxDialog":
+			show = () => eventBus.$emit('showAjaxDialogModal', {
+				title: config.title,
+				text: config.text,
+				buttons: config.buttons,
+				status: config.status,
 				port: port
 			});
 
@@ -901,7 +997,7 @@ navigator.serviceWorker.addEventListener("message", event => {
 			break;
 
 		case "showHudSettings":
-			show = () => eventBus.$emit('showSimpleMenuModal', {
+			show = () => eventBus.$emit('showAdvMenuModal', {
 				title: I18n.t("settings_title"),
 				items: config.settings,
 				port: port