+++ src/main/zapHomeFiles/hud/utils.js
@@ -130,16 +130,6 @@ var utils = (function() {
 		return hostname;
 	}
 	
-	//todo change to be sw agnostic
-	/* parses the path from a uri string */
-	function parsePathFromUrl(url) {
-		var parser = document.createElement("a");
-	
-		parser.href = url;
-		// todo: correct
-		return parser.pathname;
-	}
-	
 	/*
 	 * Return a parameter value from a uri string
 	 */
@@ -164,7 +154,7 @@ var utils = (function() {
 	/*
 	 * Initialize all of the info that will be stored in indexeddb.
 	 */
-	function initializeHUD(leftTools, rightTools) {
+	function initializeHUD(leftTools, rightTools, drawer) {
 		if (IS_DEV_MODE && leftTools.indexOf("hudErrors") < 0) {
 			// Always add the error tool in dev mode
 			leftTools.push("hudErrors");
@@ -179,6 +169,7 @@ var utils = (function() {
 		promises.push(localforage.setItem('drawer.isDrawerOpen', false));
 		// Note: in the below, "activeTab" is to be set to href, not name
 		promises.push(localforage.setItem('drawer.activeTab', '#history'));
+		promises.push(localforage.setItem('drawer', drawer));
 
 		let leftPanel = {
 			key: 'leftPanel',
@@ -652,6 +643,25 @@ var utils = (function() {
 	}
 	
 	/*
+	 * Adds the correct scheme to a url, handling the fact the ZAP could be upgrading an http url to https
+	 * Is only available in the serviceworker and Should always be used when supplying a url to the ZAP API.
+	 */
+	
+	function getUpgradedUrl(url) {
+		return localforage.getItem('upgradedDomains')
+			.then(upgradedDomains => {
+				var domain = parseDomainFromUrl(url);
+				let scheme = 'https';
+	
+				if (upgradedDomains && domain in upgradedDomains) {
+					scheme = "http";
+				}
+	
+				return scheme + "://" + domain + url.substring(url.indexOf(domain) + domain.length);
+			})
+			.catch(errorHandler)
+	}
+	/*
 	 * Log an error in a human readable way with a stack trace.
 	 */
 	function errorHandler(err) {
@@ -721,7 +731,6 @@ return {
 		parseResponseHeader: parseResponseHeader,
 		isFromTrustedOrigin: isFromTrustedOrigin,
 		parseDomainFromUrl: parseDomainFromUrl,
-		parsePathFromUrl: parsePathFromUrl,
 		getParamater: getParamater,
 		isHUDInitialized: isHUDInitialized,
 		initializeHUD: initializeHUD,
@@ -743,6 +752,7 @@ return {
 		sortToolsByPosition: sortToolsByPosition,
 		configureButtonHtml: configureButtonHtml,
 		getUpgradedDomain: getUpgradedDomain,
+		getUpgradedUrl: getUpgradedUrl,
 		errorHandler: errorHandler,
 		getZapFilePath: getZapFilePath,
 		getZapImagePath: getZapImagePath,